+++ backend/src/Service/Character.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Neucore\Service;
 
 use Neucore\Entity\CharacterNameChange;
+use Neucore\Factory\RepositoryFactory;
 
 class Character
 {
@@ -13,9 +14,15 @@ class Character
      */
     private $objectManager;
 
-    public function __construct(ObjectManager $objectManager)
+    /**
+     * @var RepositoryFactory
+     */
+    private $repositoryFactory;
+
+    public function __construct(ObjectManager $objectManager, RepositoryFactory $repositoryFactory)
     {
         $this->objectManager = $objectManager;
+        $this->repositoryFactory = $repositoryFactory;
     }
 
     /**
@@ -32,14 +39,45 @@ class Character
 
         $oldName = $character->getName();
         if ($oldName !== '' && $oldName !== $name) {
-            $characterNameChange = new CharacterNameChange();
-            $characterNameChange->setCharacter($character);
-            $characterNameChange->setOldName($oldName);
-            $characterNameChange->setChangeDate(new \DateTime());
-            $character->addCharacterNameChange($characterNameChange);
-            $this->objectManager->persist($characterNameChange);
+            $this->createNewRecord($character, $oldName);
         }
 
         $character->setName($name);
+    }
+
+    /**
+     * Checks if the name changed and adds a record for it if it is missing.
+     *
+     * Persists the new object, but does not flush the entity manager.
+     *
+     * Returns true if a new record was created, otherwise false.
+     */
+    public function addCharacterNameChange(\Neucore\Entity\Character $character, string $oldName): bool
+    {
+        if ($oldName === '' || $character->getName() === $oldName) {
+            return false;
+        }
+
+        $record = $this->repositoryFactory
+            ->getCharacterNameChangeRepository()
+            ->findOneBy(['character' => $character, 'oldName' => $oldName]);
+
+        if ($record !== null) {
+            return false;
+        }
+
+        $this->createNewRecord($character, $oldName);
+
+        return true;
+    }
+
+    private function createNewRecord(\Neucore\Entity\Character $character, string $oldName): void
+    {
+        $characterNameChange = new CharacterNameChange();
+        $characterNameChange->setCharacter($character);
+        $characterNameChange->setOldName($oldName);
+        $characterNameChange->setChangeDate(new \DateTime());
+        $character->addCharacterNameChange($characterNameChange);
+        $this->objectManager->persist($characterNameChange);
     }
 }