+++ backend/tests/Functional/Command/CheckTokensTest.php
@@ -228,7 +228,11 @@ class CheckTokensTest extends ConsoleTestCase
         $actualChars = $repositoryFactory->getCharacterRepository()->findBy([]);
         $this->assertSame(1, count($actualChars));
         $this->assertSame(3, $actualChars[0]->getId());
+        $this->assertSame('char1', $actualChars[0]->getName());
         $this->assertNull($actualChars[0]->getLastUpdate()); // token check does not change the update date
+        $nameChange = $repositoryFactory->getCharacterNameChangeRepository()->findBy([]);
+        $this->assertSame(1, count($nameChange));
+        $this->assertSame('Name', $nameChange[0]->getOldName());
     }
 
     /**